Add q_merge implementation
`q_merge` utilizes existing list merging function for mergesort. To make
use of the function, the list is first converted into a singly linked
list and then passed to the function. I iterate through all the queues
and merge every queue with the first one. Last, the prev links are
rebuilt.