Add q_delete_dup implementation
`q_delete_dup` iterates through the whole list one at a time and compare
values with the currently pinned value. If a new value is found, which
is not the same as the current pinned value, the pinned value is
assigned the new value and the pinned node is freed if it is a
duplicate. The operation will not work with empty list and list with
only one node, therefore an additional check is done before the main
traversing loop.