Add q_shuffle implementation
`q_shuffle()` is implemented using Fisher-Yates shuffle. The random
target node to be swapped is determined by a random number obtained from
standard library `rand()`. Two special cases where 1) the to be swapped
target is the current node, and 2) next_node is the target is
checked.

A helper function `list_swap()` is referred to the linux kernel
`list.h` API.