Add data preparation for sorting test
8 data patterns can be choosed from including:
- All random data
- descending data
- ascending data
- ascending, then 3 random exchanges
- ascending, add 10 random at the end
- ascending, randonly replace 1% of elements with random values
- many duplicates
- all equal elements

The implementation method can be refered to CPython `sortperf.py`
which is used to benchmark the list sort performance in CPython.
The original worst case pattern is designed to be the worstcase of the
traditional quicksort, therefore I discarded this pattern and planned to
add back more suitable worstcase scenarios to the current list
quicksort, linux mergesort and Timsort.