Add sort method test function
The test iterates through all the patterns 1~8 available currently and
size 20 to 10,000. Two sorting methods are tested, Linux `list_sort()`
and Timsort. Execution time and comparison count are the two performance
metrics used to benchmark the sorting algorithm. The output log is
printed to `sort.log` with format:
[type] [pattern] [size] [time] [comparisons]
and is ready for process.