Change random replace method
Instead of traversing the list to find the target everytime I generate a
target random number, store the to be replaced index in an array and
iterate through the array while traversing the list. This will make
larger size test more efficient.

To ensure that the random number is unique, Floyd's algorithm is used.
Quicksort is then applied to the random numbers.