Change to fixed point representation
The original value of `score` when executing mcts is tracked and
analyzed for a single run. The value mainly lies between 0 to 5 and some
occurrence of DBL_MAX. I choose 30 as the scaling factor for my fix
point representation to remain most of the precision and not sacrificing
the range of integer -2^31 ~ 2^31-1.

The four basic arithmetic operation can be done by normal operations in
combination with required shifting. Three functions are provided,
`sqrt()` for fix point square root, `fix_log2()` for base 2 logrithm and
`fix_log10()` for base 10 logrithm.

`fix_sqrt()` is done with digit by digit calculation.
`fix_log2()`'s algorithm can be refered to
http://www.claysturner.com/dsp/binarylogarithm.pdf
`fix_log10()` is calculated with `fix_log2(x)/log_2(10)`, which can be
proofed easily by log_a(b) = log_2(b) / log_2(a).