Add bot vs bot game play feature
This feature is composed of three main tasks.
1. Simulation loop
2. Key process loop
3. Screen refresh loop

The three tasks are added to the scheduling list and scheduled using
round-robin algorithm. The simulation loop simulates games
non-stoppingly until the game count exceeds the limit or `sim_active`
becomes false, which is triggered by <c-p> key press. The key process
loop reads from the raw mode terminal and only processes <c-q> and
<c-p>. The screen refreshing technique can be referred to
https://github.com/jserv/mazu-editor/tree/master.
A screen buffer is first populated with required info and the screen is
updated every cycle.